rework card structure
