+++ services/ui-src/src/components/banners/Banner.tsx
@@ -44,10 +44,9 @@ interface Props {
 
 const sx = {
   root: {
-    height: "5.25rem",
+    minHeight: "5.25rem",
     borderInlineStartWidth: "0.5rem",
     marginTop: "1.25rem",
-    marginBottom: "2.5rem",
   },
   alertIcon: {
     color: "palette.gray_darkest",
