+++ services/ui-src/src/components/index.tsx
@@ -1,6 +1,7 @@
 export { App } from "./app/App";
 export { AppRoutes } from "./app/AppRoutes";
 export { Banner } from "./banners/Banner";
+export { Card } from "./cards/Card";
 export { Footer } from "./layout/Footer";
 export { Header } from "./layout/Header";
 export { Icon } from "./icons/Icon";
@@ -10,5 +11,5 @@ export { Menu } from "./menus/Menu";
 export { MenuOption } from "./menus/MenuOption";
 export { RouterLink } from "./app/RouterLink";
 export { Table } from "./tables/Table";
-export { TemplateCard } from "./templateCards/TemplateCard";
+export { TemplateCard } from "./cards/TemplateCard";
 export { TemplateCardAccordion } from "./accordions/TemplateCardAccordion";
